+++ src/python/algorithms/WinogradScaled.py
@@ -2,13 +2,13 @@ import numpy as np
 
 def multiplicar_winograd_escalado(matriz1, matriz2, n):
     # Convertir las matrices a arrays de numpy para aprovechar operaciones vectorizadas
-    matriz1 = np.array(matriz1)
-    matriz2 = np.array(matriz2)
+    matriz1 = np.array(matriz1, dtype=int)  # Asegurarse de que matriz1 sea de tipo entero
+    matriz2 = np.array(matriz2, dtype=int)  # Asegurarse de que matriz2 sea de tipo entero
 
     # Inicializar las matrices de resultados y los factores
-    resultado = np.zeros((n, n))
-    row_factor = np.zeros(n)
-    col_factor = np.zeros(n)
+    resultado = np.zeros((n, n), dtype=int)  # Asegurarse de que el resultado sea de tipo entero
+    row_factor = np.zeros(n, dtype=int)      # Usar enteros para los factores de fila
+    col_factor = np.zeros(n, dtype=int)      # Usar enteros para los factores de columna
 
     # Precomputar los factores de las filas de matriz1
     for i in range(n):
@@ -30,4 +30,5 @@ def multiplicar_winograd_escalado(matriz1, matriz2, n):
     if n % 2 == 1:
         resultado += matriz1[:, n-1, np.newaxis] * matriz2[n-1, :]
 
+    # Convertir el resultado a una lista de listas de enteros
     return resultado.tolist()
